SEEDPipe Contents Removed
BADCODE separated from SEEDPipe contents
