+++ seedpoisoner/__init__.py
@@ -1,0 +1,1 @@
+from .SEED_Attacks.attacker import SeedPoisoner as Poisoner
